Re-structure FLI Source files
