+++ lib/fli/FliImpl.cpp
@@ -25,13 +25,12 @@
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************************************************************/
 
-#include <bitset>
 #include <vector>
 
 #include "FliImpl.h"
+#include "mti.h"
 #include "acc_vhdl.h"   // Messy :(
 #include "acc_user.h"
-
 
 extern "C" {
 
@@ -350,396 +349,6 @@ int FliImpl::deregister_callback(GpiCbHdl *gpi_hdl)
 }
 
 
-/**
- * @name    cleanup callback
- * @brief   Called while unwinding after a GPI callback
- *
- * We keep the process but de-sensitise it
- *
- * NB need a way to determine if should leave it sensitised, hmmm...
- *
- */
-int FliProcessCbHdl::cleanup_callback(void)
-{
-    if (m_sensitised) {
-        mti_Desensitize(m_proc_hdl);
-    }
-    m_sensitised = false;
-    return 0;
-}
-
-FliTimedCbHdl::FliTimedCbHdl(GpiImplInterface *impl,
-                             uint64_t time_ps) : GpiCbHdl(impl),
-                                                 FliProcessCbHdl(impl),
-                                                 m_time_ps(time_ps)
-{
-    m_proc_hdl = mti_CreateProcessWithPriority(NULL, handle_fli_callback, (void *)this, MTI_PROC_IMMEDIATE);
-}
-
-int FliTimedCbHdl::arm_callback(void)
-{
-    mti_ScheduleWakeup(m_proc_hdl, m_time_ps);
-    m_sensitised = true;
-    set_call_state(GPI_PRIMED);
-    return 0;
-}
-
-int FliTimedCbHdl::cleanup_callback(void)
-{
-    switch (get_call_state()) {
-    case GPI_PRIMED:
-        /* Issue #188: Work around for modelsim that is harmless to othes too,
-           we tag the time as delete, let it fire then do not pass up
-           */
-        LOG_DEBUG("Not removing PRIMED timer %p", m_time_ps);
-        set_call_state(GPI_DELETE);
-        return 0;
-    case GPI_CALL:
-        LOG_DEBUG("Not removing CALL timer yet %p", m_time_ps);
-        set_call_state(GPI_DELETE);
-        return 0;
-    case GPI_DELETE:
-        LOG_DEBUG("Removing Postponed DELETE timer %p", m_time_ps);
-        break;
-    default:
-        break;
-    }
-    FliProcessCbHdl::cleanup_callback();
-    FliImpl* impl = (FliImpl*)m_impl;
-    impl->cache.put_timer(this);
-    return 0;
-}
-
-int FliSignalCbHdl::arm_callback(void)
-{
-    if (NULL == m_proc_hdl) {
-        LOG_DEBUG("Creating a new process to sensitise to signal %s", mti_GetSignalName(m_sig_hdl));
-        m_proc_hdl = mti_CreateProcess(NULL, handle_fli_callback, (void *)this);
-    }
-
-    if (!m_sensitised) {
-        mti_Sensitize(m_proc_hdl, m_sig_hdl, MTI_EVENT);
-        m_sensitised = true;
-    }
-    set_call_state(GPI_PRIMED);
-    return 0;
-}
-
-int FliSimPhaseCbHdl::arm_callback(void)
-{
-    if (NULL == m_proc_hdl) {
-        LOG_DEBUG("Creating a new process to sensitise with priority %d", m_priority);
-        m_proc_hdl = mti_CreateProcessWithPriority(NULL, handle_fli_callback, (void *)this, m_priority);
-    }
-
-    if (!m_sensitised) {
-        mti_ScheduleWakeup(m_proc_hdl, 0);
-        m_sensitised = true;
-    }
-    set_call_state(GPI_PRIMED);
-    return 0;
-}
-
-FliSignalCbHdl::FliSignalCbHdl(GpiImplInterface *impl,
-                               FliSignalObjHdl *sig_hdl,
-                               unsigned int edge) : GpiCbHdl(impl),
-                                                    FliProcessCbHdl(impl),
-                                                    GpiValueCbHdl(impl, sig_hdl, edge)
-{
-    m_sig_hdl = m_signal->get_handle<mtiSignalIdT>();
-}
-
-
-GpiCbHdl *FliSignalObjHdl::value_change_cb(unsigned int edge)
-{
-    FliSignalCbHdl *cb = NULL;
-
-    switch (edge) {
-    case 1:
-        cb = &m_rising_cb;
-        break;
-    case 2:
-        cb = &m_falling_cb;
-        break;
-    case 3:
-        cb = &m_either_cb;
-        break;
-    default:
-        return NULL;
-    }
-
-    if (cb->arm_callback()) {
-        return NULL;
-    }
-
-    return (GpiValueCbHdl*)cb;
-}
-
-static const char value_enum[10] = "UX01ZWLH-";
-
-const char* FliSignalObjHdl::get_signal_value_binstr(void)
-{
-    mtiSignalIdT m_fli_hdl = get_handle<mtiSignalIdT>();
-
-    switch (m_fli_type) {
-
-        case MTI_TYPE_ENUM:
-            m_val_buff[0] = value_enum[mti_GetSignalValue(m_fli_hdl)];
-            break;
-        case MTI_TYPE_SCALAR:
-        case MTI_TYPE_PHYSICAL: {
-                std::bitset<32> value((unsigned long)mti_GetSignalValue(m_fli_hdl));
-                std::string bin_str = value.to_string<char,std::string::traits_type, std::string::allocator_type>();
-                snprintf(m_val_buff, m_val_len+1, "%s", bin_str.c_str());
-            }
-            break;
-        case MTI_TYPE_ARRAY: {
-                mti_GetArraySignalValue(m_fli_hdl, m_mti_buff);
-                if (m_val_len <= 256) {
-                    char *iter = (char*)m_mti_buff;
-                    for (int i = 0; i < m_val_len; i++ ) {
-                        m_val_buff[i] = value_enum[(int)iter[i]];
-                    }
-                } else {
-                    for (int i = 0; i < m_val_len; i++ ) {
-                        m_val_buff[i] = value_enum[m_mti_buff[i]];
-                    }
-                }
-            }
-            break;
-        default:
-            LOG_ERROR("Signal %s type %d not currently supported",
-                m_name.c_str(), m_fli_type);
-            break;
-    }
-
-    LOG_DEBUG("Retrieved \"%s\" for signal %s", m_val_buff, m_name.c_str());
-
-    return m_val_buff;
-}
-
-const char* FliSignalObjHdl::get_signal_value_str(void)
-{
-    LOG_ERROR("Getting signal value as str not currently supported");
-    return NULL;
-}
-
-double FliSignalObjHdl::get_signal_value_real(void)
-{
-    LOG_ERROR("Getting signal value as double not currently supported!");
-    return -1;
-}
-
-long FliSignalObjHdl::get_signal_value_long(void)
-{
-    LOG_ERROR("Getting signal value as long not currently supported!");
-    return -1;
-}
-
-int FliSignalObjHdl::set_signal_value(const long value)
-{
-    int rc;
-    char buff[20];
-    mtiSignalIdT m_fli_hdl = get_handle<mtiSignalIdT>();
-
-
-    snprintf(buff, 20, "16#%016X", (int)value);
-
-    rc = mti_ForceSignal(m_fli_hdl, &buff[0], 0, MTI_FORCE_DEPOSIT, -1, -1);
-
-    if (!rc) {
-        LOG_ERROR("Setting signal value failed!\n");
-    }
-    return rc-1;
-}
-
-int FliSignalObjHdl::set_signal_value(std::string &value)
-{
-    int rc;
-    mtiSignalIdT m_fli_hdl = get_handle<mtiSignalIdT>();
-
-
-    snprintf(m_val_str_buff, m_val_str_len+1, "%d'b%s", m_val_len, value.c_str());
-
-    rc = mti_ForceSignal(m_fli_hdl, &m_val_str_buff[0], 0, MTI_FORCE_DEPOSIT, -1, -1);
-    if (!rc) {
-        LOG_ERROR("Setting signal value failed!\n");
-    }
-    return rc-1;
-}
-
-int FliSignalObjHdl::set_signal_value(const double value)
-{
-    LOG_ERROR("Setting Signal via double not supported!");
-    return -1;
-}
-
-int FliSignalObjHdl::initialise(std::string &name, std::string &fq_name)
-{
-    mtiSignalIdT m_fli_hdl = get_handle<mtiSignalIdT>();
-
-    /* Pre allocte buffers on signal type basis */
-    m_fli_type = mti_GetTypeKind(mti_GetSignalType(m_fli_hdl));
-
-    switch (m_fli_type) {
-        case MTI_TYPE_ENUM:
-            m_val_len     = 1;
-            m_val_str_len = 3+m_val_len;
-            break;
-        case MTI_TYPE_SCALAR:
-        case MTI_TYPE_PHYSICAL:
-            m_val_len     = 32;
-            m_val_str_len = 4+m_val_len;
-            break;
-        case MTI_TYPE_ARRAY:
-            m_val_len     = mti_TickLength(mti_GetSignalType(m_fli_hdl));
-            m_val_str_len = snprintf(NULL, 0, "%d'b", m_val_len)+m_val_len;
-            m_mti_buff    = (mtiInt32T*)malloc(sizeof(*m_mti_buff) * m_val_len);
-            if (!m_mti_buff) {
-                LOG_CRITICAL("Unable to alloc mem for signal mti read buffer: ABORTING");
-            }
-            break;
-        default:
-            LOG_ERROR("Unable to handle onject type for %s (%d)",
-                         name.c_str(), m_fli_type);
-    }
-
-    m_val_buff = (char*)malloc(m_val_len+1);
-    if (!m_val_buff) {
-        LOG_CRITICAL("Unable to alloc mem for signal read buffer: ABORTING");
-    }
-    m_val_buff[m_val_len] = '\0';
-    m_val_str_buff = (char*)malloc(m_val_str_len+1);
-    if (!m_val_str_buff) {
-        LOG_CRITICAL("Unable to alloc mem for signal write buffer: ABORTING");
-    }
-    m_val_str_buff[m_val_str_len] = '\0';
-
-    GpiObjHdl::initialise(name, fq_name);
-
-    return 0;
-}
-
-GpiCbHdl *FliVariableObjHdl::value_change_cb(unsigned int edge)
-{
-    return NULL;
-}
-
-const char* FliVariableObjHdl::get_signal_value_binstr(void)
-{
-    mtiVariableIdT m_fli_hdl = get_handle<mtiVariableIdT>();
-
-    switch (m_fli_type) {
-
-        case MTI_TYPE_ENUM:
-            m_val_buff[0] = value_enum[mti_GetVarValue(m_fli_hdl)];
-            break;
-        case MTI_TYPE_SCALAR:
-        case MTI_TYPE_PHYSICAL: {
-                std::bitset<32> value((unsigned long)mti_GetVarValue(m_fli_hdl));
-                std::string bin_str = value.to_string<char,std::string::traits_type, std::string::allocator_type>();
-                snprintf(m_val_buff, m_val_len+1, "%s", bin_str.c_str());
-            }
-            break;
-        case MTI_TYPE_ARRAY: {
-                mti_GetArrayVarValue(m_fli_hdl, m_mti_buff);
-                if (m_val_len <= 256) {
-                    char *iter = (char*)m_mti_buff;
-                    for (int i = 0; i < m_val_len; i++ ) {
-                        m_val_buff[i] = value_enum[(int)iter[i]];
-                    }
-                } else {
-                    for (int i = 0; i < m_val_len; i++ ) {
-                        m_val_buff[i] = value_enum[m_mti_buff[i]];
-                    }
-                }
-            }
-            break;
-        default:
-            LOG_ERROR("Variable %s type %d not currently supported",
-                m_name.c_str(), m_fli_type);
-            break;
-    }
-
-    LOG_DEBUG("Retrieved \"%s\" for variable %s", m_val_buff, m_name.c_str());
-
-    return m_val_buff;
-}
-
-const char* FliVariableObjHdl::get_signal_value_str(void)
-{
-    LOG_ERROR("Getting signal value as str not currently supported");
-    return "";
-}
-
-double FliVariableObjHdl::get_signal_value_real(void)
-{
-    LOG_ERROR("Getting variable value as double not currently supported!");
-    return -1;
-}
-
-long FliVariableObjHdl::get_signal_value_long(void)
-{
-    LOG_ERROR("Getting variable value as long not currently supported!");
-    return -1;
-}
-
-int FliVariableObjHdl::set_signal_value(const long value)
-{
-    LOG_ERROR("Setting variable value not currently supported!");
-    return -1;
-}
-
-int FliVariableObjHdl::set_signal_value(std::string &value)
-{
-    LOG_ERROR("Setting variable value not currently supported!");
-    return -1;
-}
-
-int FliVariableObjHdl::set_signal_value(const double value)
-{
-    LOG_ERROR("Setting variable value not currently supported");
-    return -1;
-}
-
-int FliVariableObjHdl::initialise(std::string &name, std::string &fq_name)
-{
-    mtiVariableIdT m_fli_hdl = get_handle<mtiVariableIdT>();
-
-    /* Pre allocte buffers on signal type basis */
-    m_fli_type = mti_GetTypeKind(mti_GetVarType(m_fli_hdl));
-
-    switch (m_fli_type) {
-        case MTI_TYPE_ENUM:
-            m_val_len = 1;
-            break;
-        case MTI_TYPE_SCALAR:
-        case MTI_TYPE_PHYSICAL:
-            m_val_len = 32;
-            break;
-        case MTI_TYPE_ARRAY:
-            m_val_len  = mti_TickLength(mti_GetVarType(m_fli_hdl));
-            m_mti_buff = (mtiInt32T*)malloc(sizeof(*m_mti_buff) * m_val_len);
-            if (!m_mti_buff) {
-                LOG_CRITICAL("Unable to alloc mem for signal mti read buffer: ABORTING");
-            }
-            break;
-        default:
-            LOG_ERROR("Unable to handle object type for %s (%d)",
-                         name.c_str(), m_fli_type);
-    }
-
-    m_val_buff = (char*)malloc(m_val_len+1);
-    if (!m_val_buff) {
-        LOG_CRITICAL("Unable to alloc mem for signal read buffer: ABORTING");
-    }
-    m_val_buff[m_val_len] = '\0';
-
-    GpiObjHdl::initialise(name, fq_name);
-
-    return 0;
-}
-
 GpiIterator *FliImpl::iterate_handle(GpiObjHdl *obj_hdl, gpi_iterator_sel_t type)
 {
     /* This function should return a class derived from GpiIterator and follows it's
@@ -748,7 +357,6 @@ GpiIterator *FliImpl::iterate_handle(GpiObjHdl *obj_hdl, gpi_iterator_sel_t type
     return NULL;
 }
 
-#include <unistd.h>
 
 FliTimedCbHdl* FliTimerCache::get_timer(uint64_t time_ps)
 {
